Ignore Bitzi ticket timeouts on missing entries for explicit user retries.
When an entry is missing in the Bitzi database, the ticket received has
a 2-week expiration data, which is used to govern the time we cache the
ticket in our local database.

However, when the user explicitly attempts to re-query Bitzi, we ignore
the cached entry if it is for a missing file and the request was made
more than an hour ago (to avoid hammering, either malicious or accidental).