Discard "stable" entries we haven't heard about in last 3 months.
The probability model used to determine whether an entry is still "alive"
or not depending on its past lifetime and the time elapsed since the last
time we saw it has a long tail that can cause it to slowly discard entries
that were alive for a long time, but from which we also haven't heard for
a long time...

Therefore, put an upper limit to the time we allow between the time last
seen and the current time, so that we can discard entries that are no longer
relevant in our caches.  This limit must be high enough to not empty caches
of users who do not run the program very often, yet small enough to prevent
ever-growing caches.  The 3-month limit appears to be a good compromise.