Added ability to log extra data on assertion failure.
Added g_assert_log() to be able to log extra (possibly formatted) information
regarding the assertion failure.  This is useful in tricky situation where
a core file may not be available to probe extra data and the stack trace
would not show what is going on, either because the variables are optimized
out or because the information needs to be constructed from many data.

Propagate the error message from g_error() or the extra information logged
by g_assert_log() into the generated crash logs in case the application logs
cannot be accessed.