Don't split tail of allocated freelist blocks if remaining size too small.
This complicates the logic of inserting a block in the freelist and validating
a block size because our set of allowed size is discrete.  However it avoids
creating a large freelist of small blocks that are going to be wasted anyway
and which will consume CPU time each time we attempt coalescing.