Correctly handle shell output messages to avoid "-" in last message.
A multi-line shell output must be something like:

	100-First line
	100-Second line
	100 Last line

so that non-human clients know the last output line.

This requires buffering lines and emitting them one at a time until we reach
the end of the command, at which time we issue a final flush and know whether
this will constitute the last line of text emitted.

Also added shell_write_linef() to emit a formatted line of text and use that
to correctly emit messages in the "set" comamnd.