Added more safety (but costly) conditional assertions in zalloc()/zfree().
New routine zbelongs() is conditionally used to check that a block
address belongs to a given zone.  This is enforced at strategic times,
such as zfree() or when updating the main free list in zalloc().

Routine zvalid() makes sure a pointer in a subzone is correctly aligned
given the zone's block size.

To minimize the runtime penalty of zbelongs(), some additional bookkeeping
is necessary, albeit it was kept minimal since this is not meant to be
used in production, only when debugging problems.