Added crash_append_error() and removed unused crash_set_error_vec().
Changed the way crash_set_error() works: instead of duplicating the string
from the pre-allocated chunk, it copies it into a pre-allocated str_t object,
the same one which is used to hold assertion failure expressions.

That's because we can't have both an assertion failure and an error at the
same time, and the code was referring to a single vars->message anyway.