Reverted earlier "fix": "%015s" must not ignore the %0 directive.
The initial behaviour of str_vncatf() to emit a leading "0" sequence
was correct.  However, the GNU libc disagrees with that on Linux.

Since saying "%15s" and "%015s" indicates two different intents (the
formatted string could actually be holding digits), hence I believe that
the GNU libc is incorrect, so I restored the original behaviour.

The snprintf() of FreeBSD agrees with our initial logic as well.

To prevent undue warnings on Linux, the corresponding test in str_test()
has been flagged as "excluded".