GUESS: must forget about query if the query key was not correct.
When we refuse a GUESS query due to a wrong query key, we send back a negative
acknowledgment: a pong containing a new query key.  We then expect the remote
node to resend the same query (which will bear the same MUID) with the proper
query key this time.

Therefore, the fact that we got a query with this MUID from the sender would
make us believe that the next query with the proper query key is a duplicate
message.  To prevent this, we forget that the node sent us the message after
sending back the new query key.