MQ: don't leave flow-control condition with a partially written message.
This applies to the TCP message queue only since UDP sends messages atomically.

Staying in flow-control when the last message is partially written is useful
now that we can have more bandwidth under flow-control conditions: this allows
the queue to more quickly flush whole messages and not delay the sending of
the "hops-flow" control message which will tell our remote party that the
flow-control situation is now gone.