Don't turn off I/O favours too early for TX flow-controlled nodes.
Indeed, when dealing with TCP nodes we need to make sure the end of
flow-control condition is propagated correctly before we can resume to
normal operations.

To this end we monitor the Hops-Flow message sent to signal the end of
the flow-control condition and when it has been processed by the message
queue, we turn off the I/O favour but allocate additional bandwidth to
be able to flush the TX queue (since the compressing layer can hold data
still even though the message has been consumed from the message queue).

This should prevent remote gtk-gnutella nodes, which monitor the time
a node spends in TX flow-control, from over-estimating the actual time
and kick us out because we remain in that state for too long, apparently.