Avoid wrong "activity timeout" indications for ultra <-> ultra connections.
The time of last TX was wrongly updated when the message queue sent the
message to the TX stack, instead of taking into account the last time a
transmission physically happens on the socket (modulo TCP delaying due to
network conditions and retransmissions, naturally).

Because of the inherent buffering that occurs in the TX deflating stack, this
could lead the upper layer into thinking no transmission was happening at all,
indicating a connection hang-up, whereas the TX deflating layer was actually
engaged into a slow flushing of its buffered data, flow-controlling the upper
layer in the process.