Make sure we do not include "non-shareable" files in query hits.
These files may still be listed in the "search table" but have been de-indexed
because it was dynamically determined that they were spam, for instance.
Including them in hits would create an assertion failure later when we attempt
to build the query hit (on the grounds that their index is invalid).