Force xmalloc() and zalloc() to 8-byte alignments thanks to glib-2.30.2.
The authors of glib-2.30.2 assume that the system's malloc() will
always return pointers aligned to 8 bytes, because the GNU libc malloc()
behaves that way so of course all malloc() replacements should.

This is due to their playing with pointers in glib/gdataset.c, within
the macro G_DATALIST_GET_POINTER(), and probably at other hidden places
as well.