Monitor hops during lookups and resize/rebuild table if we have too many.
Normally we keep the hash table at a fill rate of 75% maximum, but sometimes
the dispersion of keys in the table is such that we can have more than twice
the maximum average theoretical hops during lookups (more than 8, whereas
the average is 4).

Pro-actively resize or rebuild the table in that case to keep the hash
table efficient at the expense of possibly using more memory than we could.