Randomly propagate GUESS 0.2 hosts among GUESS 0.2 servents.
Version 0.2 of the GUESS protocol extends the amount of hosts that are relayed
through query acknowledgments and helps keep GUESS queries alive.  Therefore,
it is desirable to have GUESS 0.2 hosts exchange the addresses they know
somehow.

In this first attempt, we're only including one randomly selected GUESS 0.2
host in the set we return in the IPP extension.