Smarter xmalloc sorted bucket management.
Buckets are now always sorted in such a way that the pointers we want to
allocated are put at the end, regardless of the direction of VMM allocations.

We also now manage a sorted index in the bucket, and allow xfl_lookup()
to search into the unsorted tail if it is not too large.