A bug in the sort benchmarking program led to wrong conclusions.
When sorting 4-byte quantities, we were comparing the pointers instead
of the values pointed at, completely denaturing the test since smsort()
is great when the array is initially sorted, which it appeared to be in
that case!  Since the same (wrong) routine was used to validate the results
of the sorting, the bug went unnoticed for some time.

It turns out that this implementation of smsort() is about twice slower
than qsort() in the general case, but beats down qsort() when the array
is almost sorted.

Also the internals are now sufficiently optimized to significantly blurr
the difference between smsort() and smsort_ext(), hence the loud warnings
were removed.