Fixed quicksort() to never call the comparison routine on the pivot with itself.
Now that we switched zalloc() to use xqsort(), there was an assertion failure
because subzinfo_cmp() does not expect to be called with items that are
identical!

The old code was copying the pivot item and could compare any item with that
copy, including the pivot item itself, triggering a failure.  Preventing a
call to the sort callback can be viewed as an easy optimization.