Fixed insertsort() now that we no longer scan the whole array for the minimum.
By scanning the whole array, we are ensured to find the minimim at index 0
which means we can blindly compare items because we know our minimum will
act as a sentinel: every other item is larger or equal, none can be smaller.

But since quicksort() now calls insertsort() on large arrays, we do not
wish to scan the whole range, we scan only the first few items, assuming this
will likely be the minimum since the assumption is tha the array is mostly
sorted when we are called.

As a result, we can no longer rely on the sentinel catching our reaching the
base and we need an explicit pointer comparison for that.