xgc(): fixed bug introduced in bucket garbage collecting rewrite.
Now that we don't update fl->pointers[] immediately at each removal but
defer the work until we have cleaned the whole bucket, we can't update
fl->sorted on the fly.  Rather we count the items we don't keep that were
within the sorted region to update fl->sorted only at the end.