xgc(): be smarter about coalescing of 2 blocks that need splitting.
When 2 blocks are coalesced and don't fit in a freelist bucket, we used to
ignore them.

Now the algorithm checks whether the largest of the two blocks is smaller than
the larger block that will result into the introduction of the coalesced range
into the freelist.  In that case, it can request coalescing, provided there
is room in the targeted buckets.