Added crashing modes for the VMM and xmalloc layers.
These are activated as soon as we enter the crashing mode, to limit risks
of using stale data structures when allocating and freeing memory.

In crashing mode, these allocators use the simplest possible allocation
strategy and disable freeing. Calls to xrealloc() are still possible but
also use the simple allocate-and-move strategy.