On Windows, generate a gdb-like stack trace in the crash log file.
Since there is no fork() and we do not call gdb on the executable on this
platform, generate a more complete trace, as a debugger would, collecting
symbol information using the bfd library, available by default within the
MinGW compilation environment.

Together with our native stack unwinding, this should greatly improve
the debugging of crashes on the Windows platform.