Normalized values for ctz(0) and clz(0).
These values are normally undefined, but we force normalization by defining:

	clz(0) = 32
	ctz(0) = -1

This maintains the expected assertion:

	clz(x) + ctz(x) = 31

for all values.  Of course, we could easily choose the opposite viewpoint
for 0, saying that clz(0) = -1, and this is why the value for 0 is normally
undefined.

Adapted ctz64() to understand our new convention, defining ctz64(0) = -1.