Enhanced our BFD library support on Debian systems.
On Debian, and maybe on other distributions as well, it is possible to
have a copy of the shared libraries with full debugging symbols, kept under
the /usr/lib/debug/ tree.

Whenever we have to open a shared object, look whether we do not have a
version with full symbols first.

Also changed bfd_util_compute_offset() to only consider the first section
in the file, and use page-aligned pointer comparisons to derive the offset.