Let the crash handler emit a decorated gdb-like stack if no inspector launched.
When the crash handler is invoked from a signal handler, we first emit a
simple stack trace to the log file without any decoration, to minimize the
risk of hitting corrupted data structures in the process.

Once this first stack has been emitted, we can attempt to print a decorated
stack, which will be most useful when there is no inspector launched.