Also log the faulting PC when we get a harmful signal.
On platforms where we can inspect the general machine registers from the
signal handler and determined where the PC is stored, we now log the faulting
PC along with symbolic information, when available.

Moreover, log some extended signal information if we can derive it from
the signal code value in the siginfo_t structure.