xmalloc: detect cross-thread freeing and attempt to protect execution.
Normally, the thread allocating the data should be the one freeing it, which
is rather critical when using thread-specific pools to allocate blocks.

However, it is not a model we can enforce when dealing with libraries like
GTK, which can be passed pointers to data that will become owned by it.
Unfortunately, it is hard to know beforehand when this situation is going
to occur.

We introduce an imperfect solution here whereby a thread-specific chunk list
is made lockable as soon as we detect a cross-thread freeing activity.  There
is a small window of (mis)opportunity where corruption can occur, but it is
hopefully small enough to never become a problem in practice.

A loud warning is emitted each time we first detect this cross-thread freeing
activity to be able to consider a race condition when debugging, should a crash
occur later due to a concurrency problem.