Switched halloc() to xpmalloc() instead of walloc().
It is no longer possible to base halloc() on walloc() since it is absolutely
imperative that blocks allocated by walloc() be freed by the same thread they
were allocated in, due to walloc() use of thread-specific zones for greater
performance.

Although we can ensure this in our application code, we must be careful with
blocks that can be shared among threads and for which freeing is not easy
to predict.  This is the case for most of halloc() and xmalloc() blocks.

Hence it is no longer possible to rely on walloc(), which is not much of a
problem now that xmalloc() is efficient for small blocks and can handle
freeings in a foreign thread.