halloc: during hrealloc() or hfree(), don't use size to determine type.
The block size with xpmalloc() may be larger than what was requested, hence
it is not a proper discriminent for the type of allocation that was performed.

Instead, rely on whether the block address is listed in the page table to
determine whether it was allocated via vmm_alloc() or via xpmalloc().