Removed thread suspension during error messages.
Without spinlock/mutex accounting, the potential for deadlock is huge and
this accounting is far from being cheap, so we cannot turn it on by default.
Therefore, theread suspension during error logging is dangerous.

As a workaround to the original requirement of not mixing logs whilst the
stack frame is being dumped,  the stack trace is not emitted atomically if
possible.

However, if the warning occurs in another thread during shutdown, that other
thread may not gain sufficient scheduling time to actually construct and emit
the stack frame, whereas thread suspension gave the thread all the time it
needed to be able to construct that message.

So it's still not a perfect solution to not lose stack frames during
t_carp() calls.  Also, the initial warning log and the stack frame may be
far apart in the logs due to the absence of thread suspension.