Revisited the new thread recording logic to avoid race conditions.
The old code was not behaving correctly when we were recording a new thread:
any mutex grabbed before the insertion of the new thread in the global hash
table was accounted for in the wrong thread, leading to assertion failures
later when the mutex was released by the thread.

Also changed thread_find() to make sure it now scan the thread array by
small thread ID looking for a matching thread ID when we are in the process
of creating a new thread.  This is slower because it involves a call to
thread_self() which may again do the same thing at the lower pthread level,
but this only happens when a new thread is in the critical section where it
registers itself.