Improved thread-safety of the VMM layer.
There were some corner case that were not fully thread-safe.  Most were
rather harmless (but wasting the time spent on earlier computations), one
was potentially harmful (capable of raising assertion failures in other
threads).

The harmless case is NOT fixed, but fixing it is not trivial as it could
cause a deadlock between the pmap mutex and the page cache spinlock.