PARQ: switch actively queued download to another when file has been completed.
When a downloaded is completed, we do not want to lose the actively queued
slots we have: switch them to a new resource, if any.  Some servent will not
allow the resource switching whilst waiting (which is an error) but PARQ
manages slots, so a PARQ server will allow it transparently.