When switching PARQ 1.0 downloads, switch IDs if queue position is favorable.
When a download is completed, actively queued entries are switched to another
download on the same server, if available.  When the position on the server is
more favorable, also switch the PARQ IDs so that the other download can start
sooner.  Naturally PARQ servers allow resource switching for queued entries.