Made thread quasi ID truly unique on 64-bit machines.
We used to make them fit on 32-bit quantities, but that cannot make them
totally unique among threads in a portable way.  Having them be 64-bit
quantities on 64-bit machines ensures this uniqueness and allows "same thread"
detection in thread_element_matches() when we're just one page before or
after the previously used one for the thread.