Perform xmalloc_split_setup() as soon as possible after freelist is initialized.
Large block freeing rely on the split information being computed to return the
block into the freelist, possibly split.  On Windows, this is happening earlier
than on Linux due to different C startup requirements.

Therefore, initialize static split structures as soon as the freelist is
configured and do not wait for the VMM layer to be up.