node_can_accept_protocol(): must use the Content-Type header, not Accept.
When running as ultra node we don't want to accept a G2 connection from another
ultra node (a hub in G2 parlance).  But we were wrongly looking at the Accept
header, which just mentions the possibilities, and not the actual protocol
being requested, which is indicated by the Content-Type header.

In practice, this is not going to make much of a difference as connections
with G2 nodes will likely mention "application/x-gnutella2" support in both
their Content-Type and Accept headers, but the semantic was not right.

This was an old bug introduced on 2003-01-25.