Can now add I/O callbacks triggered at beginning of scheduling timeslices.
Some I/O sources like UDP sockets (or files) are normally always available for
writing so it does not make much sense to poll/select on them.  Still, we may
want to limit the amount of I/Os done on these sources.

Therefore, it is now possible to install a "passive callback" via the new
bio_add_passive_callback() routine, which will be triggered each time the
bandwidth scheduler begins a new timeslice.