Relay GUESS query hits through semi-reliable UDP if query announced its support.
Now that queries have a flag to indicate that the emitting servent supports
semi-reliable UDP, we can remember that the GUESS client emitting the query
supports semi-reliable UDP, in case we have to route back hits to it.

Also fixed a bug whereby we were using the DHT node's message queue to relay
GUESS query hits.  Not important excepted for proper bandwidth accounting.